backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
django-cors-headers>=4.3
python-dotenv>=1.0
psycopg2-binary>=2.9
dj-database-url>=2.1
whitenoise>=6.6
gunicorn>=21.2
requests>=2.31
